refactor: consolidate XxxOptionService into each service file
